+++ scout/parse/variant/variant.py
@@ -213,9 +213,8 @@ def parse_variant(
         parsed_variant["str_pathologic_min"] = int(str_pathologic_min)
 
     ################# Add somatic info ##################
-    somatic = variant.INFO.get("SOMATIC")
-    if somatic:
-        somatic_score = variant.INFO.get("SOMATICSCORE")
+    somatic_score = variant.INFO.get("SOMATICSCORE")
+    if somatic_score:
         parsed_variant["somatic_score"] = int(somatic_score)
 
     ################# Add custom info ##################