+++ scout/server/blueprints/cases/views.py
@@ -52,7 +52,7 @@ def index():
     """Display a list of all user institutes."""
     institute_objs = user_institutes(store, current_user)
     institutes_count = (
-        (institute_obj, store.cases(collaborator=institute_obj["_id"]).count())
+        (institute_obj, sum(1 for i in store.cases(collaborator=institute_obj["_id"])))
         for institute_obj in institute_objs
         if institute_obj
     )
@@ -112,10 +112,7 @@ def matchmaker_matches(institute_id, case_name):
     data = controllers.mme_matches(case_obj, institute_obj, mme_base_url, mme_token)
     data["panel"] = panel
     if data and data.get("server_errors"):
-        flash(
-            "MatchMaker server returned error:{}".format(data["server_errors"]),
-            "danger",
-        )
+        flash("MatchMaker server returned error:{}".format(data["server_errors"]), "danger")
         return redirect(request.referrer)
     if not data:
         data = {"institute": institute_obj, "case": case_obj, "panel": panel}
@@ -260,9 +257,7 @@ def matchmaker_add(institute_id, case_name):
         store.case_mme_update(case_obj=case_obj, user_obj=user_obj, mme_subm_obj=add_result)
     flash(
         "Number of new patients in matchmaker:{0}, number of updated records:{1}, number of failed requests:{2}".format(
-            n_inserted,
-            n_updated,
-            len(add_result.get("server_responses")) - n_succes_response,
+            n_inserted, n_updated, len(add_result.get("server_responses")) - n_succes_response
         ),
         category,
     )
@@ -400,11 +395,7 @@ def pdf_case_report(institute_id, case_name):
             temp_madeline.write(case_obj["madeline_info"])
 
     html_report = render_template(
-        "cases/case_report.html",
-        institute=institute_obj,
-        case=case_obj,
-        format="pdf",
-        **data,
+        "cases/case_report.html", institute=institute_obj, case=case_obj, format="pdf", **data
     )
     return render_pdf(
         HTML(string=html_report),
@@ -471,13 +462,7 @@ def case_diagnosis(institute_id, case_name):
 
     remove = True if request.args.get("remove") == "yes" else False
     store.diagnose(
-        institute_obj,
-        case_obj,
-        user_obj,
-        link,
-        level=level,
-        omim_id=omim_id,
-        remove=remove,
+        institute_obj, case_obj, user_obj, link, level=level, omim_id=omim_id, remove=remove
     )
     return redirect(request.referrer)
 
@@ -513,11 +498,7 @@ def phenotypes(institute_id, case_name, phenotype_id=None):
             else:
                 # assume omim id
                 store.add_phenotype(
-                    institute_obj,
-                    case_obj,
-                    user_obj,
-                    case_url,
-                    omim_term=phenotype_term,
+                    institute_obj, case_obj, user_obj, case_url, omim_term=phenotype_term
                 )
         except ValueError:
             return abort(400, ("unable to add phenotype: {}".format(phenotype_term)))
@@ -594,10 +575,7 @@ def phenotypes_actions(institute_id, case_name):
         password = current_app.config["PHENOMIZER_PASSWORD"]
         diseases = controllers.hpo_diseases(username, password, hpo_ids)
         return render_template(
-            "cases/diseases.html",
-            diseases=diseases,
-            institute=institute_obj,
-            case=case_obj,
+            "cases/diseases.html", diseases=diseases, institute=institute_obj, case=case_obj
         )
 
     if action == "DELETE":
@@ -735,10 +713,7 @@ def pin_variant(institute_id, case_name, variant_id):
     variant_obj = store.variant(variant_id)
     user_obj = store.user(current_user.email)
     link = url_for(
-        "variant.variant",
-        institute_id=institute_id,
-        case_name=case_name,
-        variant_id=variant_id,
+        "variant.variant", institute_id=institute_id, case_name=case_name, variant_id=variant_id
     )
     if request.form["action"] == "ADD":
         store.pin_variant(institute_obj, case_obj, user_obj, link, variant_obj)
@@ -755,18 +730,14 @@ def mark_validation(institute_id, case_name, variant_id):
     user_obj = store.user(current_user.email)
     validate_type = request.form["type"] or None
     link = url_for(
-        "variant.variant",
-        institute_id=institute_id,
-        case_name=case_name,
-        variant_id=variant_id,
+        "variant.variant", institute_id=institute_id, case_name=case_name, variant_id=variant_id
     )
     store.validate(institute_obj, case_obj, user_obj, link, variant_obj, validate_type)
     return redirect(request.referrer or link)
 
 
 @cases_bp.route(
-    "/<institute_id>/<case_name>/<variant_id>/<partial_causative>/causative",
-    methods=["POST"],
+    "/<institute_id>/<case_name>/<variant_id>/<partial_causative>/causative", methods=["POST"]
 )
 def mark_causative(institute_id, case_name, variant_id, partial_causative=False):
     """Mark a variant as confirmed causative."""
@@ -774,23 +745,14 @@ def mark_causative(institute_id, case_name, variant_id, partial_causative=False)
     variant_obj = store.variant(variant_id)
     user_obj = store.user(current_user.email)
     link = url_for(
-        "variant.variant",
-        institute_id=institute_id,
-        case_name=case_name,
-        variant_id=variant_id,
+        "variant.variant", institute_id=institute_id, case_name=case_name, variant_id=variant_id
     )
     if request.form["action"] == "ADD":
         if "partial_causative" in request.form:
             omim_terms = request.form.getlist("omim_select")
             hpo_terms = request.form.getlist("hpo_select")
             store.mark_partial_causative(
-                institute_obj,
-                case_obj,
-                user_obj,
-                link,
-                variant_obj,
-                omim_terms,
-                hpo_terms,
+                institute_obj, case_obj, user_obj, link, variant_obj, omim_terms, hpo_terms
             )
         else:
             store.mark_causative(institute_obj, case_obj, user_obj, link, variant_obj)
@@ -976,9 +938,7 @@ def download_hpo_genes(institute_id, case_name):
         filename=f"HPO_gene_list.{institute_id}.{case_name}.{download_day}.csv",
     )
     return Response(
-        _generate_csv(",".join(csv_header), csv_lines),
-        mimetype="text/csv",
-        headers=headers,
+        _generate_csv(",".join(csv_header), csv_lines), mimetype="text/csv", headers=headers
     )