+++ scout/server/utils.py
@@ -1,6 +1,7 @@
 """Common utilities for the server code"""
 import logging
 import os
+import time
 from functools import wraps
 
 from flask import abort, flash, render_template, request