+++ CHANGELOG.md
@@ -7,19 +7,42 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [x.x.x]
 ### Added
 - Load and show SpliceAI annotations
+- Institute-level phenotype models with sub-panels containing HPO and OMIM terms
+- Runnable Docker demo
+- Docker image build and push github action
+### Fixed
+- Update dismissed variant status when variant dismissed key is missing
+- Breakpoint two IGV button now shows correct chromosome when different from bp1
+- Missing font lib in Docker image causing the PDF report download page to crash
+- Sentieon Manta calls lack Somaticscore - load anyway
+### Changed
+- Make matching causative and managed variants foldable on case page
+- Remove calls to PyMongo functions marked as deprecated in backend and frontend(as of version 3.7).
+- Improved `scout update individual` command
+
+## [4.23]
+### Added
+- Save custom IGV track settings
+- Show a flash message with clear info about non-valid genes when gene panel creation fails
+- CNV report link in cancer case side navigation
+- Return to comment section after editing, deleting or submitting a comment
+- Managed variants
+- MT vs 14 chromosome mean coverage stats if Scout is connected to Chanjo
 ### Fixed
 - missing `vcf_cancer_sv` and `vcf_cancer_sv_research` to manual.
-
+- Split ClinVar multiple clnsig values (slash-separated) and strip them of underscore for annotations without accession number
+- Timeout of `All SNVs and INDELs` page when no valid gene is provided in the search
+- Round CADD (MIPv9)
+- Missing default panel value
+- Invisible other causatives lines when other causatives lack gene symbols
 ### Changed
 - Do not freeze mkdocs-material to version 4.6.1
 - Remove pre-commit dependency
-
 
 ## [4.22]
 ### Added
 - Editable cases comments
 - Editable variants comments
-- CNV report link in cancer case side navigation
 ### Fixed
 - Empty variant activity panel
 - STRs variants popover